memoize initialized breakpoints
